+++ R/zzz.R
@@ -7,6 +7,23 @@
 # Created: 2011-12-11
 # */
 
+
+# backports
+# R < 3.6.0
+if (getRversion() < "3.6.0") {
+  errorCondition = function(message, ..., class = NULL, call = NULL) {
+    obj <- list(message = as.character(message), call = call, ...)
+    class(obj) = c(class, "error", "condition")
+    obj
+  }
+  
+  warningCondition = function(message, ..., class = NULL, call = NULL) {
+    obj <- list(message = as.character(message), call = call, ...)
+    class(obj) = c(class, "warning", "condition")
+    obj
+  }
+}
+  
 # nocov start
 .onUnload = function(libpath) {
   library.dynam.unload("bit64", libpath)
@@ -216,3 +233,50 @@ deprecate_exported_s3_methods(
   unipos.integer64
 )
 # nocov end
+
+
+# The call stack is searched for a given sequence of function names. If the last element of function_names is found, 
+# we try to match as many elements in function_names with the function names in the call stack. The sequence must be 
+# adhered to. The complete call of the function name of the last match is returned. If no match exists, the top call
+# is returned. It is also possible to change the function name of the matched return value by providing its new name
+# with name_to_display.
+# Examples: 
+# * call stack: [A, B, C, D, E]; function_names = c("C", "D") returns C
+# * call stack: [A, B, C, D, E]; function_names = c("E", "D") returns D
+# * call stack: [A, B, C, D, E]; function_names = c("E", "X") returns A
+choose_sys_call = function(function_names, name_to_display=NULL) {
+  calls = sys.calls()
+  if (length(calls) == 1L || length(function_names) == 0L) return(calls[[1L]])
+  # find last occurrence of last name in function_names
+  function_names_rev = rev(as.character(function_names))
+  for (sel in rev(seq_along(calls))) {
+    el = calls[[sel]]
+    if (!is.function(el[[1L]]) && rev(as.character(el[[1L]]))[1L] == function_names_rev[1L]) break
+  }
+  # now check further backwards to match as far as possible
+  for (fn in function_names_rev[-1L]) {
+    if (sel == 1L) break
+    el = calls[[sel - 1L]]
+    if (is.function(el[[1L]]) || rev(as.character(el[[1L]]))[[1L]] != fn) break
+    sel = sel - 1L
+  }
+  ret = calls[[sel]]
+  if (!is.null(name_to_display))
+    ret[[1L]] = as.name(name_to_display)
+  ret
+}
+
+withCallingHandlers_and_choose_call = function(expr, function_names, name_to_display=NULL) {
+  wch = substitute(
+    withCallingHandlers(expr, error=error, warning=warning),
+    list(
+      expr = sys.call()[[2L]],
+      error = function(e) stop(errorCondition(e$message, call=choose_sys_call(function_names, name_to_display))),
+      warning = function(w) {
+        warning(warningCondition(w$message, call=choose_sys_call(function_names, name_to_display)))
+        invokeRestart("muffleWarning")
+      }
+    )
+  )
+  eval(wch, envir=parent.frame())
+}